Add assertions and parameter checking.
Public methods need to throw exceptions when parameters are invalid.
Nonpublic methods need to check program state at runtime.

Let's,
* add parameter checking for most of the public methods and
  constructors
* add assertions to identify bugs

+++ src/main/java/duke/command/DeleteAllCommand.java
@@ -7,7 +7,10 @@ import duke.ui.Ui;
 
 public class DeleteAllCommand implements ICommand {
     @Override
-    public void execute(TaskList taskList, Ui ui, Storage storage) throws DukeException {
+    public void execute(TaskList taskList, Ui ui, Storage storage) throws DukeException, IllegalArgumentException {
+        if (taskList == null || ui == null || storage == null) {
+            throw new IllegalArgumentException("One of the parameters is null.");
+        }
         taskList.deleteAll();
         storage.removeAll();
         Ui.printRemoveAll();
